non_iter_update merged with Niter2
