+++ .gitignore
@@ -3,3 +3,4 @@ figs/*
 House/*
 Dataset/robot/images/*
 Dataset/drone/images/*
+*.zip
